Subzero: Fix a performance regression in the register allocator.
The register allocator does not need to be considering rematerializable variables at all.  When it does, there tends to be a big performance cost because the live range of a rematerializable variable tends to be large, leading to lots of expensive overlap computations against register allocation candidates.

BUG= none
R=jpp@chromium.org, sehr@chromium.org

Review URL: https://codereview.chromium.org/1450233002 .