Subzero: Add "szbuild_spec2k.py --run" option.
This makes it easier to do bisection debugging.

Note that this could be used to simplify "make -f Makefile.standalone check-spec", but I chose not to do that because shellcmd() buffers all of the stdout until completion, which is not a great user experience (at least for this user).

BUG= none
R=sehr@chromium.org

Review URL: https://codereview.chromium.org/1472833006 .