fix doxygen brief in subzero header files
There is a consistent mistake in most of the file level subzero doxygen
comments which stems from a problem with brief.

Basically, there are two types of documentation in Doxygen, inline and
structural documentation. File level documentation is always structural
because inline documentation must either come before or after the entity
and for a file that would be impossible.

When you have structure documentation, the brief and detailed is
handled differently and you have to explicity put the \brief qualifier in.

We were only doing this for a few files so when you looked the doxygen
for the source directory, the description was blank for almost all of them.
If you build docs with this patch and look at the "files" and src
directory, you will see all the brief descriptions now.

In addition I deleted the text "This file ..." from the
beginning of all the file desriptions because it's redundant when you look at the doxygen output and adds
no information but takes up space in the descripton.
BUG=
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1488913002 .