Improve error handling in the ARM integrated assembler.
Up to now, all error handling in the ARM integrated assembler was
handled by forcing a text fixup. This CL tries to minimize the use of
fixup's to only be applied when there is an unimplemented form of an
instruction. All other cases now generate fatal error messages.

This CL should make it easier to determine what instructions still need
to be extended.

BUG= https://bugs.chromium.org/p/nativeclient/issues/detail?id=4334
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1495093002 .