add predefines to doxygen
this is added for both standard doxygen and clang assisted
doxygen (which cannot be turned on yet without some additions to the third-party toolchain but can be used locally).

the predefines here are the defaults when subzero is built.

not having these creates the bigger issue when clang
assisted parsing is enabled but should be there always.

BUG=
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1533963002 .

Patch from Reed Kotler <rkotlerimgtec@gmail.com>.