add include files from llvm used by subzero into doxygen
this fixes a lot of problems with the doxygen. for example
with this, now doxygen correctly understands the cl::opt
variables in ClFlags.cpp as variables, whereas previously
it thought they were functions.

it's possible to make the llvm includes more selective but
this is a good first cut and better than including all
of llvm includes.

BUG=
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1536773003 .

Patch from Reed Kotler <rkotlerimgtec@gmail.com>.