doxygenize IceClFlags.cpp
This is a first cut at this. More can be done and much of this is just moving what is already in the help for the
commands into Doxygen comments. The documentation can be
expanded to better describe the role of the various
command line options.

BUG=
R=stichnot@chromium.org

Review URL: https://codereview.chromium.org/1547743002 .

Patch from Reed Kotler <rkotlerimgtec@gmail.com>.