Subzero: Fix -build-atts option.
If the default --target option is not actually included in the Subzero build (via llvm/Config/SZTargets.def), then "pnacl-sz --build-atts" will fail.  This is because the attribute printing is done after the GlobalContext is created, which does some amount of Target initialization.

The fix is to move the attribute printing to a point after the flags are parsed and the output streams are created, but before the GlobalContext is created.  This basically disables --build-atts in the browser build, but that should be OK.

BUG= none
R=kschimpf@google.com

Review URL: https://codereview.chromium.org/1584923002 .