Subzero: Change the szbuild.py --force default.
With the --force option, the behavior of its default setting is dangerous.  Specifically, it doesn't retranslate by default when you change the szbuild command-line arguments.  This leads to debugging frustration.

Therefore, we change the default setting to be "dumb" and always retranslate.

For bisection debugging, the user will quickly realize that they want to add "--force=0".

BUG= none
R=jpp@chromium.org

Review URL: https://codereview.chromium.org/1608323002 .