Update patch set 3
Patch Set 3:

Right. Fortunately lvalue() is not an issue since uniforms are read-only. And while they can be used as a relative index, that only goes for scalars. So even if a single element of a uniform matrix is used as an index, we're first fully reading the matrix and then selecting a single row and column. So the .rel applies to a temporary, not a uniform.

Please verify.

Patch-set: 3