Update patch set 4
Patch Set 4:

We need to this to be active on all builds.

We need to name the blocks with their start addresses to make Android stack traces intelligible. Android stack traces assume that all code will be in shared libraries. The backtrace is printing in the form:

<hex_offset> <library_name>

That works well for shared libraries because it save the reader time when looking up the code in the library. However, it doesn't work well for code sitting on the malloc heap: there is no way to convert the relative offset to the absolute position in memory. Things get even worse if there are multiple blocks of the same size.

I'd actually like to use memory mapped files for the blocks at some point so we can get the JIT'd code as part of the tombstone. If we do that we can pull this code back out (the blocks will automatically be tagged with the file path). However, that's actually a more complicated change.

Patch-set: 4
Reviewer: Greg Hartman <5200@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=0