Subzero: Fix -timing-focus .
The problem is that the default timer is renamed to be the current function name, instead of "Total across all functions", but this causes an assertion failure during the final merge of thread-local timers into the global timers, because the names don't match.

The solution is to temporarily rename the timer to be the function name, until the timer is dumped, and then restore it to its original value.

BUG= none
R=jpp@chromium.org

Review URL: https://codereview.chromium.org/1867473002 .