Update patch set 1
Patch Set 1:

dEQP-GLES2 on Android has valid.local_variable_hides_function_parameter, whereas dEQP-GLES3 has invalid.local_variable_hides_function_parameter. So it follows section 4.2.2 of the specifications. However, it's not part of the mustpass set, so some drivers must be disagreeing on this.

I'll work on a patch to support both behaviors with a single grammar.

Patch-set: 1